+++ lib/ultimate_turbo_modal.rb
@@ -2,12 +2,9 @@
 
 require_relative "ultimate_turbo_modal/version"
 require "phlex/deferred_render_with_main_content"
-require "ultimate_turbo_modal/native_sheet_config"
 require "ultimate_turbo_modal/helpers/inline_action_renderer"
 require "ultimate_turbo_modal/helpers/modal_action_renderer"
 require "ultimate_turbo_modal/helpers/action_builder"
-require "ultimate_turbo_modal/helpers/native_overflow_item"
-require "ultimate_turbo_modal/helpers/native_split_action_renderer"
 require "ultimate_turbo_modal/configuration"
 require "ultimate_turbo_modal/railtie"
 require "ultimate_turbo_modal/base"