+++ test/tmp/dummy_app/app/javascript/controllers/native_sheet_controller.js
@@ -1,7 +1,0 @@
-import { Controller } from "@hotwired/stimulus"
-
-export default class extends Controller {
-  dismiss() {
-    window.history.back()
-  }
-}